DAS-1901: Addressing PR comments. Switching to a mutation and renaming token to authHeader
