+++ src/utils/parseRequestedFields.js
@@ -113,8 +113,8 @@ export const parseRequestedFields = (parsedInfo, keyMap, conceptName) => {
   }
 
   if (name === 'collection' || name === 'collections') {
-    // If a user has requested granules, subscriptions, relatedCollections, duplicateCollections,
-    // generateVariableDrafts or publishVariableDrafts from
+    // If a user has requested granules, subscriptions, relatedCollections,
+    // duplicateCollections or generateVariableDrafts from
     // within a collection request the resolver will pull the conceptId and provide
     // it to the granules request but if a user doesn't explicity ask for the
     // collection concept id we need to request it
@@ -125,7 +125,6 @@ export const parseRequestedFields = (parsedInfo, keyMap, conceptName) => {
         || requestedFields.includes('relatedCollections')
         || requestedFields.includes('duplicateCollections')
         || requestedFields.includes('generateVariableDrafts')
-        || requestedFields.includes('publishVariableDrafts')
       )
        && !requestedFields.includes('conceptId')) {
       requestedFields.push('conceptId')
